feat(get): validate response content type when using `--accept`
