+++ src/main/java/me/itzg/helpers/get/OutputToDirectoryHandler.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.hc.core5.http.ClassicHttpResponse;
 import org.apache.hc.core5.http.HttpException;
@@ -14,6 +15,8 @@ class OutputToDirectoryHandler implements OutputResponseHandler {
   private final Path directory;
   final FilenameExtractor filenameExtractor;
   private final boolean logProgressEach;
+  private String expectedContentType;
+  private ContentTypeValidator contentTypeValidator;
 
   public OutputToDirectoryHandler(Path directory,
       LatchingUrisInterceptor interceptor, boolean logProgressEach) {
@@ -23,8 +26,16 @@ class OutputToDirectoryHandler implements OutputResponseHandler {
   }
 
   @Override
+  public void setExpectedContentTypes(List<String> contentTypes) {
+    this.contentTypeValidator = new ContentTypeValidator(contentTypes);
+  }
+
+  @Override
   public Path handleResponse(ClassicHttpResponse response) throws HttpException, IOException {
     LoggingResponseHandler.logResponse(log, response);
+    if (contentTypeValidator != null) {
+      contentTypeValidator.validate(response);
+    }
 
     final String filename = filenameExtractor.extract(response);
