+++ src/main/java/me/itzg/helpers/get/SaveToFileHandler.java
@@ -4,7 +4,9 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.List;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.hc.core5.http.ClassicHttpResponse;
 import org.apache.hc.core5.http.HttpEntity;
 
 @Slf4j
@@ -12,10 +14,24 @@ public class SaveToFileHandler extends LoggingResponseHandler<Path> implements O
 
   private final Path outputFile;
   private final boolean logProgressEach;
+  private ContentTypeValidator contentTypeValidator;
 
   public SaveToFileHandler(Path outputFile, boolean logProgressEach) {
     this.outputFile = outputFile;
     this.logProgressEach = logProgressEach;
+  }
+
+  @Override
+  public void setExpectedContentTypes(List<String> contentTypes) {
+    contentTypeValidator = new ContentTypeValidator(contentTypes);
+  }
+
+  @Override
+  public Path handleResponse(ClassicHttpResponse response) throws IOException {
+    if (contentTypeValidator != null) {
+      contentTypeValidator.validate(response);
+    }
+    return super.handleResponse(response);
   }
 
   @Override
